requirements.txt
flask
pybase62
redis
pymongo
pytest
